+++ src/app/handlers/drag/interfaces.ts
@@ -1,0 +1,4 @@
+export interface ICuiElementDetector {
+    detect(x: Number, y: Number): [number, Element];
+    setElements(elements: Element[]): void;
+}